Move tasks API tests
